+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to this library are documented in this file.
 
 ### Bug Fixes
 
+- Fix thread-safely of embedded `pymemcache` client (#148).
 - Implement ATmosphere/bluesky media upload again (#144).
 - Removed unnecessary `pwd` (UNIX only) module usage (#142).